chore: remove Uuid distinction
